# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /google/status:
    get:
      operationId: getGoogleStatus
      tags: [google]
      summary: Get Google connection status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleStatus"
  /google/auth:
    get:
      operationId: getGoogleAuthUrl
      tags: [google]
      summary: Get Google OAuth authorization URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Auth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleAuthUrl"
  /google/disconnect:
    post:
      operationId: disconnectGoogle
      tags: [google]
      summary: Disconnect Google account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
  /google/location:
    put:
      operationId: setGoogleLocation
      tags: [google]
      summary: Set Google Business location
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleLocationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  locationId:
                    type: string
                required:
                  - locationId
  /google/locations:
    get:
      operationId: getGoogleLocations
      tags: [google]
      summary: List Google Business locations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Locations list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleLocationsResponse"
  /google/sync:
    post:
      operationId: syncGoogleProfile
      tags: [google]
      summary: Sync selected fields to Google Business Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSyncRequest"
      responses:
        "200":
          description: Sync results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleSyncResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GoogleStatus:
      type: object
      properties:
        connected:
          type: boolean
        email:
          type: string
        locationId:
          type: string
        lastSyncHours:
          type: string
        lastSyncDescription:
          type: string
        lastSyncPhotos:
          type: string
        lastSyncPosts:
          type: string
        lastSyncWebsiteUrl:
          type: string
        statusHours:
          type: string
        statusDescription:
          type: string
        statusPhotos:
          type: string
        statusPosts:
          type: string
        statusWebsiteUrl:
          type: string
      required:
        - connected
    GoogleAuthUrl:
      type: object
      properties:
        authUrl:
          type: string
      required:
        - authUrl
    GoogleLocationInput:
      type: object
      properties:
        locationId:
          type: string
        accountId:
          type: string
      required:
        - locationId
        - accountId
    GoogleLocation:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
        name:
          type: string
        accountName:
          type: string
      required:
        - id
        - accountId
        - name
        - accountName
    GoogleLocationsResponse:
      type: object
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/GoogleLocation"
      required:
        - locations
    GoogleSyncRequest:
      type: object
      properties:
        fields:
          type: array
          items:
            type: string
      required:
        - fields
    GoogleSyncResult:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
        syncedAt:
          type: string
      required:
        - success
    GoogleSyncResponse:
      type: object
      properties:
        results:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/GoogleSyncResult"
      required:
        - results
...